requirements.txt
google-generativeai>=0.7.0
gspread>=6.0.0
google-auth>=2.28.0
streamlit>=1.32.0
pandas>=2.0.0
